pyproject.toml
[project]
name = "ctc-filter"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "mkdir700", email = "mkdir700@gmail.com" }]
dependencies = [
    "python-okx>=0.2.9",
    "python-dotenv>=1.0.1",
    "TA-Lib>=0.4.29",
    "pandas>=2.2.2",
    "typer>=0.12.3",
    "pydantic>=2.7.2",
    "pydantic-settings>=2.2.1",
    "python-telegram-bot>=21.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.scripts]
ctc-filter = "ctc_filter.cli:app"


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ipykernel>=6.29.4",
    "mplfinance>=0.12.10b0",
    "tqdm>=4.66.4",
    "ipywidgets>=8.1.3",
]
